include cred stack failure reason in error
